refactor: extract prompt suggestions menu to separate component
