+++ components/Chat/ChatInput.tsx
@@ -10,9 +10,6 @@ import {
   IconMicrophone,
   IconPlayerStopFilled,
   IconMicrophone2,
-  IconBulb,
-  IconChevronLeft,
-  IconChevronRight,
 } from '@tabler/icons-react';
 import {
   KeyboardEvent,
@@ -36,6 +33,8 @@ import { compressImage, getWorkflowName } from '@/utils/app/helper';
 import { Message } from '@/types/chat';
 
 import HomeContext from '@/pages/api/home/home.context';
+
+import { PromptSuggestions } from './PromptSuggestions';
 
 interface Props {
   onSend: (message: Message) => void;
@@ -78,10 +77,7 @@ export const ChatInput = ({
     useState('');
   const [isRecording, setIsRecording] = useState(false);
   const recognitionRef = useRef(null);
-  const [showPromptGuide, setShowPromptGuide] = useState(false);
-  const [selectedCategory, setSelectedCategory] = useState<string | null>(null);
   const [promptSuggestions, setPromptSuggestions] = useState<Record<string, string[]> | null>(null);
-  const promptGuideRef = useRef<HTMLButtonElement>(null);
 
   const triggerFileUpload = () => {
     fileInputRef?.current.click();
@@ -354,6 +350,13 @@ export const ChatInput = ({
     }
   }, [isRecording]);
 
+  const handlePromptSelect = (prompt: string) => {
+    setContent(prompt);
+    if (textareaRef && textareaRef.current) {
+      textareaRef.current.focus();
+    }
+  };
+
   const loadPromptSuggestions = async () => {
     const customSuggestions = await loadContentFile<Record<string, string[]>>('promptSuggestions.json', true);
     if (customSuggestions) {
@@ -367,52 +370,12 @@ export const ChatInput = ({
       loadPromptSuggestions();
     }
 
-    // Cleanup: stop speech recognition on unmount
     return () => {
       if (recognitionRef.current) {
         recognitionRef.current.stop();
       }
     };
   }, []);
-
-  // Handle clicks outside prompt suggestions to close it
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        promptGuideRef.current &&
-        !promptGuideRef.current.contains(event.target as Node) &&
-        !(event.target as Element).closest('.prompt-guide-menu')
-      ) {
-        setShowPromptGuide(false);
-        setSelectedCategory(null);
-      }
-    };
-
-    if (showPromptGuide) {
-      document.addEventListener('mousedown', handleClickOutside);
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [showPromptGuide]);
-
-  const handlePromptSelect = (prompt: string) => {
-    setContent(prompt);
-    setShowPromptGuide(false);
-    setSelectedCategory(null);
-    if (textareaRef && textareaRef.current) {
-      textareaRef.current.focus();
-    }
-  };
-
-  const handleCategorySelect = (category: string) => {
-    setSelectedCategory(category);
-  };
-
-  const handleBackToCategories = () => {
-    setSelectedCategory(null);
-  };
 
   return (
     <div
@@ -529,75 +492,15 @@ export const ChatInput = ({
               <IconMicrophone size={18} />
             )}
           </button>
-          
+
           {promptSuggestions && (
-            <button
-              ref={promptGuideRef}
-              onClick={() => {
-                setShowPromptGuide(!showPromptGuide);
-                if (showPromptGuide) {
-                  setSelectedCategory(null);
-                }
-              }}
-              className={`absolute left-10 top-2 rounded-sm p-[5px] text-neutral-800 opacity-60 dark:bg-opacity-50 dark:text-neutral-100 ${
-                messageIsStreaming
-                  ? 'text-neutral-400'
-                  : 'hover:text-[#76b900] dark:hover:text-neutral-200'
-              }`}
-              disabled={messageIsStreaming}
-            >
-              <IconBulb size={18} />
-            </button>
+            <PromptSuggestions
+              promptSuggestions={promptSuggestions}
+              messageIsStreaming={messageIsStreaming}
+              onPromptSelect={handlePromptSelect}
+            />
           )}
-          
-          {showPromptGuide && promptSuggestions && (
-            <div className="prompt-guide-menu absolute left-2 bottom-14 w-96 max-h-[500px] overflow-y-auto bg-white dark:bg-[#40414F] border border-neutral-200 dark:border-gray-700 rounded-lg shadow-lg z-50">
-              <div className="p-4">
-                <div className="flex items-center justify-between mb-3">
-                  {selectedCategory && (
-                    <button
-                      onClick={handleBackToCategories}
-                      className="flex items-center text-sm text-neutral-600 dark:text-neutral-400 hover:text-neutral-800 dark:hover:text-neutral-200"
-                    >
-                      <IconChevronLeft size={16} />
-                      <span>Back</span>
-                    </button>
-                  )}
-                  <h3 className="text-sm font-semibold text-neutral-800 dark:text-neutral-100">
-                    {selectedCategory || 'Prompt Suggestions'}
-                  </h3>
-                  <div className="w-12"></div>
-                </div>
-                
-                {!selectedCategory ? (
-                  <div className="space-y-2">
-                    {Object.keys(promptSuggestions).map((category) => (
-                      <button
-                        key={category}
-                        onClick={() => handleCategorySelect(category)}
-                        className="w-full flex items-center justify-between py-2 text-sm text-neutral-700 dark:text-neutral-300 hover:bg-neutral-100 dark:hover:bg-gray-700 rounded transition-colors"
-                      >
-                        <span>{category}</span>
-                        <IconChevronRight size={16} />
-                      </button>
-                    ))}
-                  </div>
-                ) : (
-                  <div className="space-y-2">
-                    {promptSuggestions[selectedCategory]?.map((prompt, index) => (
-                      <button
-                        key={index}
-                        onClick={() => handlePromptSelect(prompt)}
-                        className="w-full text-left py-2 text-sm text-neutral-700 dark:text-neutral-300 hover:bg-neutral-100 dark:hover:bg-gray-700 rounded transition-colors"
-                      >
-                        {prompt}
-                      </button>
-                    ))}
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
+
           <button
             className="absolute right-2 top-2 rounded-sm p-1 text-neutral-800 opacity-60 hover:bg-neutral-200 hover:text-neutral-900 dark:bg-opacity-50 dark:text-neutral-100 dark:hover:text-neutral-200"
             onClick={handleSend}
